fix use-after-free of query results in sql insert
previously references to fields of the query result Q
could remain in the action buffer when no references to Q
itself were alive. this would cause Q to get gc'd and the
references to become garbage. fix by allocating Q on m3
heap instead of luajit heap, so it stays alive until
the end of the branch.